+++ library/Morloc/CodeGenerator/Grammars/Translator/R.hs
@@ -36,7 +36,7 @@ translate srcs es = do
     (unique . mapMaybe srcPath $ srcs)
 
   -- diagnostics
-  liftIO . putDoc . vsep $ map pretty es
+  debugLog (vsep (map pretty es))
 
   -- translate each manifold tree, rooted on a call from nexus or another pool
   mDocs <- mapM translateManifold es
@@ -50,6 +50,11 @@ translate srcs es = do
     , scriptCode = "." :/ File "pool.R" (Code . render $ code)
     , scriptMake = [SysExe outfile]
     }
+
+debugLog :: Doc ann -> MorlocMonad ()
+debugLog d = do
+  verbosity <- MM.gets stateVerbosity
+  when (verbosity > 0) $ (liftIO . putDoc) d
 
 letNamer :: Int -> MDoc 
 letNamer i = "a" <> viaShow i